requirements/production.txt
flask
flask-sqlalchemy
flask_bcrypt
flask_wtf
flask_login
Flask-APScheduler
APScheduler
Flask-Babel